#497 Moved RateLimitExceededException to its own module
