+++ src/backend_api/app/core/rate_limit/custom_rate_limiter.py
@@ -11,7 +11,7 @@ from asyncio_redis_rate_limit.compat import (
 )
 from typing_extensions import final
 
-from app.exceptions.model_exceptions.plot_exception import RateLimitExceededError
+from app.exceptions.model_exceptions.rate_limit_exception import RateLimitExceededError
 
 #: These aliases makes our code more readable.
 _Seconds: TypeAlias = int
